[TEST] Corrected RNG initialization

+++ src/augs/misc/randomization_declaration.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstdint>
 
-using rng_seed_type = unsigned;
+using rng_seed_type = uint64_t;
 struct randomization;
